refactor: [FE] QuizReplySection의 0번 라운드에 관련된 로직을 별도 hook으로 분리
